Remove optics and added thin lens file.
